+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Interop/AzureMapsInterop.cs
@@ -36,6 +36,11 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop
         IAzureMapsAnimations Animations { get; }
 
         /// <summary>
+        /// <inheritdoc cref="IAzureMapsAtlas"/>
+        /// </summary>
+        IAzureMapsAtlas Atlas { get; }
+
+        /// <summary>
         /// <inheritdoc cref="IAzureMapsCommon"/>
         /// </summary>
         IAzureMapsCommon Common { get; }
@@ -140,6 +145,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop
             dotNetRef = DotNetObjectReference.Create(component);
 
             Animations = new AzAnimations(moduleTask);
+            Atlas = new AzAtlas(moduleTask);
             Common = new AzCommon(moduleTask);
             Controls = new AzControls(moduleTask);
             Configuration = new AzConfiguration(moduleTask);
@@ -155,6 +161,8 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop
         }
 
         public IAzureMapsAnimations Animations { get; }
+
+        public IAzureMapsAtlas Atlas { get; }
 
         public IAzureMapsCommon Common { get; }